Set deletion_topic and scene_topic to fix warnings
The Scene3D plugin can be configured with scene_topic and
deletion_topic names via xml. Information can be published
to these topics about models and lights to be added or
deleted from the scene. Delphyne is not yet using this
functionality, but we add unused topic names to the xml
to silence a console warning.